requirements.txt
# Core runtime
google-genai>=1.0.0
youtube-transcript-api>=0.6.2
pillow>=10.0.0

# Optional for Level B screenshots
playwright>=1.40.0

# System tools (install separately): ffmpeg, yt-dlp, tesseract
